Use security manager instead of bootloader
Grub2 is acting as a bootloader but also as the security interface that
will talk with LUKS, the TMP2 and other subsystems to produce the
encryption.

The alternative can be use systemd components (like systemd-cryptenroll)
to act as a different security interface, with independence of the
bootloader used.

Remove the references that conflates the security realm with the
bootloader selected.

Signed-off-by: Alberto Planas <aplanas@suse.com>